trial for adding constraint using map stream

+++ src/main/java/ru/ifmo/ctddev/gmwcs/Main.java
@@ -25,7 +25,7 @@ public class Main {
         optionParser.acceptsAll(asList("n", "nodes"), "Node list file").withRequiredArg().required();
         optionParser.acceptsAll(asList("e", "edges"), "Edge list file").withRequiredArg().required();
         optionParser.acceptsAll(asList("nm", "number"), "Preserved Nodes").withRequiredArg()
-                .ofType(Integer.class).defaultsTo(10);
+                .ofType(Double.class).defaultsTo(10.0);
         optionParser.accepts("root", "Root node").withRequiredArg();
         optionParser.acceptsAll(asList("m", "threads"), "Number of threads").withRequiredArg()
                 .ofType(Integer.class).defaultsTo(1);
@@ -80,7 +80,7 @@ public class Main {
         TimeLimit tl = new TimeLimit(timelimit <= 0 ? Double.POSITIVE_INFINITY : timelimit);
         double rsh = (Double) optionSet.valueOf("r");
         double ush = (Double) optionSet.valueOf("u");
-        int max_num_nodes = (Integer) optionSet.valueOf("nm");
+        double max_num_nodes = (Double) optionSet.valueOf("nm");
         TimeLimit biggestTL = tl.subLimit(1.0 - ush);
         int threadsNum = (Integer) optionSet.valueOf("threads");
         File nodeFile = new File((String) optionSet.valueOf("nodes"));
